requirements.txt
altgraph==0.17.4
barcode==1.0.4
Levenshtein==0.25.1
numpy==2.0.1
packaging==24.1
pandas==2.2.2
pillow==10.4.0
pyinstaller==6.9.0
pyinstaller-hooks-contrib==2024.7
python-barcode==0.15.1
python-dateutil==2.9.0.post0
python-Levenshtein==0.25.1
pytz==2024.1
pyzbar==0.1.9
rapidfuzz==3.9.6
six==1.16.0
tzdata==2024.1
